Try `usePaymentActivityDateRangePresets` custom hook to clarify
